# wwwroot/openapi.json
{
"openapi": "3.0.3",
"info": {
  "title": "TripTastic API",
  "description": "Travel API for searching and booking flights and hotels. Supported destinations: New York (JFK), Los Angeles (LAX), Chicago (ORD), Dallas (DFW), Denver (DEN), San Francisco (SFO), Seattle (SEA), Miami (MIA), Boston (BOS), Atlanta (ATL), London (LHR), Paris (CDG), Frankfurt (FRA), Tokyo (NRT), Sydney (SYD). IMPORTANT: Call /api/System/current-date first to get today's date.",
  "version": "1.0.0"
},
"servers": [{"url": "https://triptastic.azurewebsites.net"}],
"paths": {
...
  "/api/TripPlanner/plan": {
    "get": {
      "operationId": "planTrip",
      "summary": "Get flight and hotel options for a round-trip",
      "description": "Returns available outbound flights (origin to destination on departure date), return flights (destination to origin on return date), and hotels at the destination. Use the IDs from the results to add items to cart.",
      "parameters": [
        {"name": "origin", "in": "query", "required": true, "schema": {"type": "string"}, "description": "Origin airport code (e.g., SEA, JFK, LAX)"},
        {"name": "destination", "in": "query", "required": true, "schema": {"type": "string"}, "description": "Destination airport code (e.g., LAX, CDG, NRT)"},
        {"name": "departureDate", "in": "query", "required": true, "schema": {"type": "string", "format": "date"}, "description": "Departure date (YYYY-MM-DD)"},
        {"name": "returnDate", "in": "query", "required": true, "schema": {"type": "string", "format": "date"}, "description": "Return date (YYYY-MM-DD)"},
        {"name": "travelers", "in": "query", "schema": {"type": "integer", "default": 1}, "description": "Number of travelers"}
      ],
      "responses": {"200": {"description": "Success", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TripPlan"}}}}}
    }
  },
...
  "/api/Cart/trips/{tripId}": {
    "get": {
      "operationId": "getTrip",
      "summary": "Get a specific booked trip",
      "description": "Returns details of a specific booked trip by ID.",
      "parameters": [
        {"name": "tripId", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}, "description": "The trip ID"}
      ],
      "responses": {
        "200": {"description": "Success", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/BookedTripResponse"}}}},
        "404": {"description": "Trip not found"}
      }
    }
  },
...
},
"components": {
"schemas": {
...
  "Flight": {
    "type": "object",
    "properties": {
      "id": {"type": "string", "format": "uuid"},
      "airline": {"type": "string"},
      "flightNumber": {"type": "string"},
      "origin": {"type": "string"},
      "destination": {"type": "string"},
      "departureTime": {"type": "string", "format": "date-time"},
      "arrivalTime": {"type": "string", "format": "date-time"},
      "price": {"type": "number"}
    }
  },
  "Hotel": {
    "type": "object",
    "properties": {
      "id": {"type": "string", "format": "uuid"},
      "name": {"type": "string"},
      "location": {"type": "string"},
      "starRating": {"type": "integer"},
      "pricePerNight": {"type": "number"},
      "amenities": {"type": "array", "items": {"type": "string"}}
    }
  },
    "HotelAvailability": {
      "type": "object",
      "properties": {
        "hotel": {"$ref": "#/components/schemas/Hotel"},
        "totalPrice": {"type": "number"},
        "nights": {"type": "integer"}
      }
    },
    "TripPlan": {
      "type": "object",
      "description": "Available flights and hotels for a trip. Use the IDs to add items to cart.",
      "properties": {
        "origin": {"type": "string"},
        "destination": {"type": "string"},
        "departureDate": {"type": "string", "format": "date"},
        "returnDate": {"type": "string", "format": "date"},
        "nights": {"type": "integer"},
        "travelers": {"type": "integer"},
        "outboundFlights": {"type": "array", "items": {"$ref": "#/components/schemas/Flight"}, "description": "Flights from origin to destination on departure date"},
        "returnFlights": {"type": "array", "items": {"$ref": "#/components/schemas/Flight"}, "description": "Flights from destination to origin on return date"},
        "hotels": {"type": "array", "items": {"$ref": "#/components/schemas/HotelAvailability"}, "description": "Hotels at the destination"}
      }
    },
...
    "BookedTripResponse": {
      "type": "object",
      "properties": {
        "id": {"type": "string", "format": "uuid"},
        "confirmationCode": {"type": "string"},
        "status": {"type": "string"},
        "bookedAt": {"type": "string", "format": "date-time"},
        "totalPrice": {"type": "number"},
        "flights": {"type": "array", "items": {"$ref": "#/components/schemas/BookedFlightInfo"}},
        "hotels": {"type": "array", "items": {"$ref": "#/components/schemas/BookedHotelInfo"}}
      }
    },
    "BookedFlightInfo": {
      "type": "object",
      "properties": {
        "id": {"type": "string", "format": "uuid"},
        "confirmationCode": {"type": "string"},
        "airline": {"type": "string"},
        "flightNumber": {"type": "string"},
        "origin": {"type": "string"},
        "destination": {"type": "string"},
        "departureTime": {"type": "string", "format": "date-time"},
        "arrivalTime": {"type": "string", "format": "date-time"},
        "passengers": {"type": "integer"},
        "pricePerSeat": {"type": "number"},
        "totalPrice": {"type": "number"}
      }
    },
    "BookedHotelInfo": {
      "type": "object",
      "properties": {
        "id": {"type": "string", "format": "uuid"},
        "confirmationCode": {"type": "string"},
        "hotelName": {"type": "string"},
        "location": {"type": "string"},
        "starRating": {"type": "integer"},
        "checkInDate": {"type": "string", "format": "date"},
        "checkOutDate": {"type": "string", "format": "date"},
        "nights": {"type": "integer"},
        "rooms": {"type": "integer"},
        "guests": {"type": "integer"},
        "pricePerNight": {"type": "number"},
        "totalPrice": {"type": "number"}
      }
    },
...
  }
}
}